requirements.txt
# simulator
matplotlib
pyserial
# visualizer
py5==0.9.1a1
